requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-nested-admin==4.0.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
